create table query corrected

+++ server_and_client/database.sql
@@ -2,6 +2,6 @@ CREATE TABLE scores_test(
     id SERIAL PRIMARY KEY,
     name VARCHAR(100) NOT NULL,
     score INTEGER NOT NULL CHECK(score >= 0),
-    time_taken INTEGER NOT NULL CHECK (time_taken >= 0)
+    time_taken DOUBLE PRECISION NOT NULL CHECK (time_taken >= 0)
 );
